Cluster Torture test: check inner backup created on cluster node full sync + minor sanity fixes
